requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML/AI
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.3.1
scikit-learn==1.4.0
faiss-cpu==1.7.4
scipy==1.11.4

# Data Processing
pandas==2.1.4
numpy==1.26.3
yfinance==0.2.35

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2

# Caching
redis==5.0.1

# MLflow
mlflow==2.9.2

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# HTTP
requests==2.31.0
httpx==0.26.0

# Configuration
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.0

# XAI & Explainability (NEW)
shap==0.44.0
lime==0.2.0.1

# Dashboard & Visualization (NEW)
streamlit==1.29.0
plotly==5.18.0
seaborn==0.13.1
matplotlib==3.8.2
rich==13.7.0

# NLP & Fine-tuning (NEW)
datasets==2.16.1
tokenizers==0.15.0
accelerate==0.25.0
tensorboard==2.15.1

# Advanced ML (NEW)
xgboost==2.0.3
optuna==3.5.0
